third_party/WebKit/LayoutTests/css3/device-adapt/opera/descriptor-values-001-expected.txt


FAIL CSS Test: @viewport accept both percentage and number values for zoom factors assert_equals: expected 3 but got 1
